% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.gss-package}
\alias{mlr3learners.gss}
\alias{mlr3learners.gss-package}
\title{mlr3learners.gss: Learners from the {gss} package for 'mlr3'}
\description{
Adds learners from the {gss} package to 'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
